+++ sdk/eventhubs/azure-messaging-eventhubs/src/test/java/com/azure/messaging/eventhubs/EventHubBufferedProducerAsyncClientIntegrationTest.java
@@ -168,11 +168,15 @@ public class EventHubBufferedProducerAsyncClientIntegrationTest extends Integrat
 
         final Random randomInterval = new Random(10);
         final Map<String, List<String>> expectedPartitionIdsMap = new HashMap<>();
+        final Map<String, String> expectedPartitionKeysMap = new HashMap<>();
         final PartitionResolver resolver = new PartitionResolver();
 
         final List<Mono<Integer>> publishEventMono = IntStream.range(0, numberOfEvents).mapToObj(index -> {
             final String partitionKey = "partition-" + index;
             final EventData eventData = new EventData(partitionKey);
+            eventData.setMessageId(String.valueOf(index));
+            expectedPartitionKeysMap.put(eventData.getMessageId(), partitionKey);
+
             final SendOptions sendOptions = new SendOptions().setPartitionKey(partitionKey);
             final int delay = randomInterval.nextInt(20);
 
@@ -207,6 +211,10 @@ public class EventHubBufferedProducerAsyncClientIntegrationTest extends Integrat
         for (SendBatchSucceededContext context : succeededContexts) {
             final List<String> expected = expectedPartitionIdsMap.get(context.getPartitionId());
             assertNotNull(expected, "Did not find any expected for partitionId: " + context.getPartitionId());
+
+            for (EventData eventData : context.getEvents()) {
+                assertEquals(expectedPartitionKeysMap.get(eventData.getMessageId()), eventData.getPartitionKey());
+            }
 
             context.getEvents().forEach(eventData -> {
                 final boolean success = expected.removeIf(key -> key.equals(eventData.getBodyAsString()));